Mock - timers and core scheduling
Spawn a second thread for setup() & loop()
Implement Core task switching routines
Implement basic support for timers
(tnx to Sming code for the idea of keeping timestamps in ETSTimer struct)

+++ libraries/Ticker/src/Ticker.cpp
@@ -74,7 +74,7 @@ void Ticker::_attach(Ticker::Milliseconds milliseconds, bool repeat)
 
     // whenever duration excedes this limit, make timer repeatable N times
     // in case it is really repeatable, it will reset itself and continue as usual
-    size_t total = 0;
+    uint32_t total = 0;
     if (milliseconds > DurationMax) {
         total = 1;
         while (milliseconds > DurationMax) {
